+++ UsercentricsChallenge/app/src/main/java/com/jpp/usercentricschallenge/ui/home/HomeViewState.kt
@@ -2,5 +2,5 @@ package com.jpp.usercentricschallenge.ui.home
 
 data class HomeViewState(
     val collectButtonEnabled: Boolean = false,
-    val cost: Int = 0,
+    val cost: Double = 0.0,
 )
